timberclaw/backend/requirements.txt
# TimberClaw Builder business API (Django) — separate from OpenHands Python stack (M0-01 scaffold).
Django>=5.0,<6
djangorestframework>=3.15,<4
httpx>=0.27,<0.29
gunicorn>=22.0
psycopg[binary]>=3.2
dj-database-url>=2.1
pytest>=8.0
pytest-django>=4.8
